+++ contracts/src/utils/math/storage.rs
@@ -1,5 +1,5 @@
 //! Simple math operations missing in `stylus_sdk::storage`.
-use alloy_primitives::{ruint::UintTryFrom, Uint};
+use alloy_primitives::Uint;
 use stylus_sdk::storage::StorageUint;
 
 /// Adds value and assign the result to `self`, ignoring overflow.
@@ -8,13 +8,11 @@ pub(crate) trait AddAssignUnchecked<T> {
     fn add_assign_unchecked(&mut self, rhs: T);
 }
 
-impl<T, const B: usize, const L: usize> AddAssignUnchecked<T>
+impl<const B: usize, const L: usize> AddAssignUnchecked<Uint<B, L>>
     for StorageUint<B, L>
-where
-    Uint<B, L>: UintTryFrom<T>,
 {
-    fn add_assign_unchecked(&mut self, rhs: T) {
-        let new_balance = self.get() + Uint::<B, L>::from(rhs);
+    fn add_assign_unchecked(&mut self, rhs: Uint<B, L>) {
+        let new_balance = self.get() + rhs;
         self.set(new_balance);
     }
 }
@@ -25,13 +23,11 @@ pub(crate) trait SubAssignUnchecked<T> {
     fn sub_assign_unchecked(&mut self, rhs: T);
 }
 
-impl<T, const B: usize, const L: usize> SubAssignUnchecked<T>
+impl<const B: usize, const L: usize> SubAssignUnchecked<Uint<B, L>>
     for StorageUint<B, L>
-where
-    Uint<B, L>: UintTryFrom<T>,
 {
-    fn sub_assign_unchecked(&mut self, rhs: T) {
-        let new_balance = self.get() - Uint::<B, L>::from(rhs);
+    fn sub_assign_unchecked(&mut self, rhs: Uint<B, L>) {
+        let new_balance = self.get() - rhs;
         self.set(new_balance);
     }
 }