+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Refactor how `AddAssignUnchecked` and `SubAssignUnchecked` traits are used in `Erc20`, `Erc721`, and `Erc721Consecutive`. #467
+- Use `AddAssignUnchecked` and `SubAssignUnchecked` in `erc20::_update` calculations. #467
 
 ### Changed (Breaking)